Rename ast_traits module to traits
