+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_new_expr/get_source_info.rs
@@ -1,5 +1,5 @@
 use super::AzleNewExpr;
-use crate::ts_ast::{ast_traits::GetSourceInfo, source_map::GetSourceFileInfo};
+use crate::ts_ast::{source_map::GetSourceFileInfo, traits::GetSourceInfo};
 
 impl GetSourceInfo for AzleNewExpr<'_> {
     fn get_source(&self) -> String {
