+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_fn_or_constructor_type/azle_fn_type/function_and_method_type_helper_methods.rs
@@ -1,7 +1,7 @@
 use swc_ecma_ast::{TsFnParam, TsType, TsTypeAnn};
 
 use super::AzleFnType;
-use crate::ts_ast::{ast_traits::GetTsType, FunctionAndMethodTypeHelperMethods};
+use crate::ts_ast::{traits::GetTsType, FunctionAndMethodTypeHelperMethods};
 
 impl FunctionAndMethodTypeHelperMethods for AzleFnType<'_> {
     fn get_ts_fn_params(&self) -> Vec<TsFnParam> {
