+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_fn_or_constructor_type/errors.rs
@@ -3,7 +3,7 @@ use swc_ecma_ast::TsFnOrConstructorType;
 
 use crate::{
     errors::ErrorMessage,
-    ts_ast::{ast_traits::GetSpan, source_map::GetSourceFileInfo},
+    ts_ast::{source_map::GetSourceFileInfo, traits::GetSpan},
 };
 
 pub(super) fn constructor_not_supported_error(
