+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_keyword_type/errors.rs
@@ -3,7 +3,7 @@ use swc_ecma_ast::TsKeywordTypeKind;
 use super::AzleKeywordType;
 use crate::{
     errors::{ErrorMessage, Suggestion},
-    ts_ast::{ast_traits::GetSourceInfo, source_map::GetSourceFileInfo, GetSourceText},
+    ts_ast::{source_map::GetSourceFileInfo, traits::GetSourceInfo, GetSourceText},
 };
 
 impl AzleKeywordType<'_> {
