+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_tuple_type/get_source_info.rs
@@ -1,5 +1,5 @@
 use super::AzleTupleType;
-use crate::ts_ast::{ast_traits::GetSourceInfo, source_map::GetSourceFileInfo};
+use crate::ts_ast::{source_map::GetSourceFileInfo, traits::GetSourceInfo};
 
 impl GetSourceInfo for AzleTupleType<'_> {
     fn get_source(&self) -> String {
