+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/azle_type_element/azle_method_signature/errors.rs
@@ -1,4 +1,4 @@
-use crate::{errors::ErrorMessage, ts_ast::ast_traits::GetSourceInfo};
+use crate::{errors::ErrorMessage, ts_ast::traits::GetSourceInfo};
 
 use super::AzleMethodSignature;
