+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/azle_type_element/azle_property_signature/get_source_info.rs
@@ -1,5 +1,5 @@
 use super::AzlePropertySignature;
-use crate::ts_ast::{ast_traits::GetSourceInfo, source_map::GetSourceFileInfo};
+use crate::ts_ast::{source_map::GetSourceFileInfo, traits::GetSourceInfo};
 
 impl GetSourceInfo for AzlePropertySignature<'_> {
     fn get_source(&self) -> String {
