+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/azle_type_element/azle_property_signature/mod.rs
@@ -1,7 +1,7 @@
 use swc_common::SourceMap;
 use swc_ecma_ast::TsPropertySignature;
 
-use crate::ts_ast::{ast_traits::GetTsType, azle_type::AzleType, GetName};
+use crate::ts_ast::{azle_type::AzleType, traits::GetTsType, GetName};
 use cdk_framework::act::node::{
     data_type::{record, variant},
     DataType,
