+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/azle_type_element/errors.rs
@@ -5,8 +5,8 @@ use super::AzleTypeElement;
 use crate::{
     errors::{ErrorMessage, Suggestion},
     ts_ast::{
-        ast_traits::{GetSourceInfo, GetSpan, TypeToString},
         source_map::GetSourceFileInfo,
+        traits::{GetSourceInfo, GetSpan, TypeToString},
     },
 };
