+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/azle_type_element/get_source_info.rs
@@ -1,5 +1,5 @@
 use super::AzleTypeElement;
-use crate::ts_ast::ast_traits::GetSourceInfo;
+use crate::ts_ast::traits::GetSourceInfo;
 
 impl GetSourceInfo for AzleTypeElement<'_> {
     fn get_source(&self) -> String {
