+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/azle_type_element/get_span.rs
@@ -1,4 +1,4 @@
-use crate::ts_ast::ast_traits::GetSpan;
+use crate::ts_ast::traits::GetSpan;
 
 use super::AzleTypeElement;
