+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_lit/azle_type_element/type_to_string.rs
@@ -1,5 +1,5 @@
 use super::AzleTypeElement;
-use crate::ts_ast::ast_traits::TypeToString;
+use crate::ts_ast::traits::TypeToString;
 
 impl TypeToString for AzleTypeElement<'_> {
     fn type_to_string(&self) -> String {
