+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/azle_type/azle_type_ref/to_data_type.rs
@@ -13,7 +13,7 @@ use super::AzleTypeRef;
 use crate::{
     generators::func,
     ts_ast::{
-        ast_traits::GetTsType, azle_type::AzleType, AzleFnOrConstructorType,
+        azle_type::AzleType, traits::GetTsType, AzleFnOrConstructorType,
         FunctionAndMethodTypeHelperMethods, GetName,
     },
 };
