+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/class_decl/get_dependencies.rs
@@ -1,7 +1,7 @@
 use std::collections::{HashMap, HashSet};
 use swc_ecma_ast::{ClassDecl, ClassMember};
 
-use crate::ts_ast::{ast_traits::GetDependencies, source_map::SourceMapped, AzleTypeAliasDecl};
+use crate::ts_ast::{source_map::SourceMapped, traits::GetDependencies, AzleTypeAliasDecl};
 
 impl GetDependencies for Vec<SourceMapped<'_, ClassDecl>> {
     fn get_dependent_types(
