+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/mod.rs
@@ -1,15 +1,12 @@
-use ast_traits::GetDependencies;
-use ast_traits::GetSourceText;
-use ast_traits::GetTsType;
 use azle_functions_and_methods::FunctionAndMethodTypeHelperMethods;
-use azle_type::AzleArrayType;
-use azle_type::AzleFnOrConstructorType;
+use azle_type::{AzleArrayType, AzleFnOrConstructorType};
+use traits::{GetDependencies, GetSourceText, GetTsType};
 
-pub use ast_traits::GetName;
 pub use azle_fn_decl::AzleFnDecl;
 pub use azle_new_expr::AzleNewExpr;
 pub use azle_program::AzleProgram;
 pub use azle_type_alias_decls::AzleTypeAliasDecl;
+pub use traits::GetName;
 pub use ts_ast::TsAst;
 
 mod azle_fn_decl;
@@ -26,7 +23,6 @@ mod ts_type;
 mod ts_type_alias_decl;
 mod ts_type_ann;
 
-pub mod ast_traits;
 pub mod azle_program;
 pub mod azle_type;
 pub mod azle_type_alias_decls;
@@ -34,3 +30,4 @@ pub mod class_decl;
 pub mod class_prop;
 pub mod module_item;
 pub mod param;
+pub mod traits;
