+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/traits/get_name.rs
