+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_fn_or_constructor_type.rs
@@ -1,6 +1,6 @@
 use swc_ecma_ast::TsFnOrConstructorType;
 
-use crate::ts_ast::ast_traits::GetSpan;
+use crate::ts_ast::traits::GetSpan;
 
 impl GetSpan for TsFnOrConstructorType {
     fn get_span(&self) -> swc_common::Span {
