+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type_ann.rs
@@ -1,6 +1,6 @@
 use swc_ecma_ast::TsTypeAnn;
 
-use super::ast_traits::GetTsType;
+use super::traits::GetTsType;
 
 impl GetTsType for TsTypeAnn {
     fn get_ts_type(&self) -> swc_ecma_ast::TsType {
